+++ perft.h
@@ -1,3 +1,3 @@
 #include "types.h"
 
-int perft(position& pos, int tiefe, int drucktiefe);
+uint64_t perft(position& pos, int tiefe, int drucktiefe);
